implement nip-02
