+++ packages/nostr/test/dm.ts
@@ -1,7 +1,8 @@
-import { createDirectMessage, EventKind, signEvent } from "../src/event"
+import { EventKind, signEvent } from "../src/event"
 import { parsePublicKey } from "../src/crypto"
 import assert from "assert"
 import { setup } from "./setup"
+import { createDirectMessage } from "../src/event/direct-message"
 
 describe("dm", () => {
   const message = "for your eyes only"
@@ -112,12 +113,11 @@ describe("dm", () => {
 
         subscriber.on("eose", async () => {
           // TODO No signEvent, do something more convenient
-          const event = await signEvent(
-            await createDirectMessage({
+          const event = await createDirectMessage(
+            {
               message,
               recipient: recipientPubkey,
-              priv: publisherSecret,
-            }),
+            },
             publisherSecret
           )
           publisher.publish(event)
